REFACTOR!: changing the database to link frames with video
The database now has new table videos and links the frames  with the video. The PK for frames is not frame_number anymore but frame_id as it goes on and the frame_number resets based on the video, so the database does not have to be cleared on each start of the app.
TODO: need to distinguish yolo and bytetrack so the frames does not overwrite itself?